complete frontend

+++ backend/accident-analytics-api/src/main/java/com/sujan/accident/analytics/repository/unfall/AccidentCategoryRepository.java
@@ -3,5 +3,5 @@ package com.sujan.accident.analytics.repository.unfall;
 import com.sujan.accident.analytics.model.unfall.AccidentCategory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AccidentCategoryRepository extends JpaRepository<AccidentCategory, Long> {
+public interface AccidentCategoryRepository extends JpaRepository<AccidentCategory, Integer> {
 }
