+++ backend/accident-analytics-api/src/main/java/com/sujan/accident/analytics/repository/unfall/AccidentKindRepository.java
@@ -3,5 +3,5 @@ package com.sujan.accident.analytics.repository.unfall;
 import com.sujan.accident.analytics.model.unfall.AccidentKind;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AccidentKindRepository extends JpaRepository<AccidentKind, Long> {
+public interface AccidentKindRepository extends JpaRepository<AccidentKind, Integer> {
 }
