+++ backend/accident-analytics-api/src/main/java/com/sujan/accident/analytics/repository/unfall/AccidentTypeRepository.java
@@ -3,5 +3,5 @@ package com.sujan.accident.analytics.repository.unfall;
 import com.sujan.accident.analytics.model.unfall.AccidentType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AccidentTypeRepository extends JpaRepository<AccidentType, Long> {
+public interface AccidentTypeRepository extends JpaRepository<AccidentType, Integer> {
 }
