+++ backend/accident-analytics-api/src/main/java/com/sujan/accident/analytics/repository/unfall/LightConditionRepository.java
@@ -3,5 +3,5 @@ package com.sujan.accident.analytics.repository.unfall;
 import com.sujan.accident.analytics.model.unfall.LightCondition;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LightConditionRepository extends JpaRepository<LightCondition, Long> {
+public interface LightConditionRepository extends JpaRepository<LightCondition, Integer> {
 }
