+++ backend/accident-analytics-api/src/main/java/com/sujan/accident/analytics/repository/unfall/PlausibilityLevelRepository.java
@@ -3,5 +3,5 @@ package com.sujan.accident.analytics.repository.unfall;
 import com.sujan.accident.analytics.model.unfall.PlausibilityLevel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PlausibilityLevelRepository extends JpaRepository<PlausibilityLevel, Long> {
+public interface PlausibilityLevelRepository extends JpaRepository<PlausibilityLevel, Integer> {
 }
