+++ backend/accident-analytics-api/src/main/java/com/sujan/accident/analytics/repository/unfall/RoadConditionRepository.java
@@ -3,5 +3,5 @@ package com.sujan.accident.analytics.repository.unfall;
 import com.sujan.accident.analytics.model.unfall.RoadCondition;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RoadConditionRepository extends JpaRepository<RoadCondition, Long> {
+public interface RoadConditionRepository extends JpaRepository<RoadCondition, Integer> {
 }
